+++ WorldOfTheThreeKingdoms/PluginInterface/PluginInterface.cs
@@ -283,10 +283,10 @@ namespace PluginInterface
 
     public interface IPersonPortrait : IBasePlugin
     {
-        bool HasPortrait(int id);
-        System.Drawing.Image GetImage(int id);
-        Texture2D GetPortrait(int id);
-        Texture2D GetSmallPortrait(int id);
+        bool HasPortrait(float id);
+        System.Drawing.Image GetImage(float id);
+        Texture2D GetPortrait(float id);
+        Texture2D GetSmallPortrait(float id);
         void SetGraphicsDevice(GraphicsDevice device);
     }
